Added Year Parameter to Find-TheMovieDBTVShowID
Added the year parameter to the Find-TheMovieDBTVShowID function because
after moving the word "the" end of of the folder name, ie. "Witcher, The",
it would have finding a result in the search.  Adding the Year parameter
was one more thing that could be done to help improve the search results.