Fixed Problems in  FolderStructure Script
Added error handling around if the folder already existed to stop the
script instead of continue on the script and returning a lot of errors.
Also the script didn't understand how to create sub folder structures
so I added code around there so that it would but the subfolders in the
correct parent folders instead of the root of the main.