External Subtitle Files Support
Added code supporting external subtitle files.
It will either grab the subtitle file with the same name as the episode
and rename it to the new episode name or it will find a folder with the
episode name in it and rename the files found in there to the episode
name and add a counter at the end of the file name starting with 2.
Needed to add ignore message for vscode because it was complaining about
the $count variable being defined in the begin block of the foreach loop
and not in the process part of the loop.  Currently located on Line 433.