Bug Fixes & Specified Video File Types
Figured out that [IO.Path]::GetInvalidFileNameChars()) does not return
anything on Linux.  I did not want to have to specify video file types,
but figured out that I needed to since dealing with subtitle files too.
Also the regex that remove all characters doesn't work as expected in
PowerShell.  I would like to fix this in the future.